feat: 컴피존 세팅 > 컴피존 이름 View 구현 및 Cell 분리
